+++ src/dev/icon-buttons/secondary-actions-menu-button/index.js
@@ -1,3 +1,3 @@
 // @flow
-export { FavoriteIconButton } from "./favorite-icon-button";
+
 export { SecondaryActionsMenuButton } from "./secondary-actions-menu-button";